Provide for lazy commit of page-map meta-data in TCMalloc
Define a variant of the flat-map that was historically defined as
PageMap1, which allocated about 4MB in an array to map all possible
pages to their status as part of a span. This version (under Windows
only) uses VirtualAlloc to effectively reserve the space, and then
only commit the space when it is necessary to ensure it is
available.

The bulk of the code is almost directly modeled on PageMap1, which is
also defined in pagemap.h.

bug=30010
r=willchan
Review URL: http://codereview.chromium.org/460155

git-svn-id: http://src.chromium.org/svn/trunk/src/third_party/tcmalloc/chromium@34422 4ff67af0-8c30-449e-8e8b-ad334ec8d88c