Fix a bug in gperftools' logging system on android
When deep memory profiler is enabled on android, the default priority of the log was ANDROID_LOG_UNKNOWN, which caused the log can't be seen.
This patch added a VERBOSE severity level, which made the log with default severity can be logged properly with a ANDROID_LOG_VERBOSE priority.

R=dmikurube@chromium.org
BUG=

Review URL: https://codereview.chromium.org/209873002

git-svn-id: http://src.chromium.org/svn/trunk/src/third_party/tcmalloc/chromium@261254 4ff67af0-8c30-449e-8e8b-ad334ec8d88c